Ajout de ask_to_delete dans le model pour que les utilisateurs peuvent demander la suppression de leur ticket

+++ ticket/views/tickets.py
@@ -356,8 +356,13 @@ def delete_ticket(request, id):
     :param id:
     :return:
     """
-    Follow.objects.filter(ticket_id=id).delete()
-    Tickets.objects.filter(id=id).delete()
+    if request.user.is_staff:
+        Follow.objects.filter(ticket_id=id).delete()
+        Tickets.objects.filter(id=id).delete()
+    else:
+        ticket = Tickets.objects.get(pk=id)
+        ticket.ask_to_delete = 1
+        ticket.save()
     return redirect('/')
